feat: add unit test, create job sync exchange rate, add document code, add swagger ui

+++ src/main/java/com/tpi/currency/dto/currency/Currency.java
@@ -7,4 +7,4 @@ public record Currency(
     String symbol,
     String rate,
     String description,
-    @JsonProperty("rate_float") double rateFloat) {}
+    @JsonProperty("rate_float") Float rateFloat) {}
